+++ ext/awsauth/examples/defaultconfig/defaultconfig.go
@@ -4,10 +4,11 @@
 // not use this file except in compliance with the License. You may obtain
 // a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
 
-package awsauth_test
+package main
 
 import (
 	"context"
+	"log"
 
 	"github.com/aws/aws-sdk-go-v2/config"
 	"go.mongodb.org/mongo-driver/ext/awsauth"
@@ -15,21 +16,21 @@ import (
 	"go.mongodb.org/mongo-driver/v2/mongo/options"
 )
 
-func Example_defaultConfig() {
+func main() {
 	cfg, err := config.LoadDefaultConfig(context.TODO())
 	if err != nil {
-		panic(err)
+		log.Fatalf("failed to load AWS config: %v", err)
 	}
+
 	awsCredentialProvider := awsauth.NewCredentialsProvider(cfg.Credentials)
 	credential := options.Credential{
 		AuthMechanism:          "MONGODB-AWS",
 		AWSCredentialsProvider: awsCredentialProvider,
 	}
-	client, err := mongo.Connect(
-		options.Client().SetAuth(credential),
-	)
+
+	client, err := mongo.Connect(options.Client().SetAuth(credential))
 	if err != nil {
-		panic(err)
+		log.Fatalf("failed to connect: %v", err)
 	}
-	_ = client
+	defer func() { _ = client.Disconnect(context.Background()) }()
 }
