+++ go.work
@@ -3,12 +3,12 @@ go 1.25.0
 use (
 	.
 	./examples
+	./ext/awsauth
+	./ext/awsauth/examples
 	./internal/cmd/benchmark
 	./internal/cmd/check-fmt
 	./internal/cmd/faas/awslambda/mongodb
+	./internal/test/awsauth
 	./internal/test/compilecheck
 	./internal/test/goleak
-	./ext/awsauth
-	./ext/awsauth/examples
-	./ext/awsauth/test
 )
